moved packages into pkg dir
